Fix Makefile not finding GIO libs on certain systems
When I tried to compile `mpv-mpris` on my Gentoo machine, it couldn't
find the directory where GIO was stored, so to fix this issue, I have
modified the Makefile to supply `--cflags` to `pkg-config`, which
supplies the path to look for the GIO libraries.

Because of the success of this fix, I am submitting it as a PR to
resolve build problems that may occur in the future without this fix.