Only send Seeked signal after mpv seek event
The MPV_PLAYBACK_RESTART event may occur on playback events that are not
seek events. In particular, this event fires when tracks change.

According to the MPRIS Seeked event specification, this is discouraged.

> This signal does not need to be emitted when playback starts or when
> the track changes, unless the track is starting at an unexpected
> position.

It may additionally cause a race condition where the Seeked signal may
be emitted before the track changes.

Instead use the MPV_EVENT_SEEK to determine when a seek begins and only
send the mpris Seeked signal once the seek has completed.