Do not queue on initial name request
The name lost handler runs when the name cannot be acquired because
another application is using it, presumably another instance of
mpv-mpris. By default, this puts mpv-mpris into a queue for the name.
When the other name becomes available, mpv-mpris will then acquire it
and own both names which is a mistake.

Add the DO_NOT_QUEUE name flag to disable the queuing behavior to fix
this issue.