Add a method to install the plugin system-wide for all users
This will be helpful for distributions packaging the plugin.

Use a symlink instead of installing a binary in the /etc directory,
since that is meant to be configuration files only, unfortunately
mpv does not load plugins from anywhere in the /usr directory tree.

Support the DESTDIR variable for overriding the rootfs location
and support overriding the PREFIX and the plugin directories.