Discard the playerctl stderr when listing players
This fixes a test failure when using glib 2.73.3.

playerctl with glib 2.73.3 emits a deprecation warning and that is
merged into the player list and the comparison with 'mpv' then fails.

When mpv is not present, the reason for that still needs to be printed,
so re-run the command when it failed or did not detect mpv is the player.