Fix idle->playing state transition
When mpv is started with the `--idle=yes` CLI flag, MPV_EVENT_IDLE fires
almost immediately at process start. This causes mpv-mpris to mark
`PlaybackStatus` as `STOPPED`.

When a file/URL is subsequently loaded into mpv and starts playing,
mpv-mpris continues to believe that playback is stopped.

Turns out, MPV_EVENT_IDLE was designed only for a state transition into
idle state, and is itself deprecated, with the "idle-active" property
suggested as a proper substitute. This property changes correctly both
when entering as well as exiting idle state.

Curiously, the "pause" property starts out as false in `--idle=yes`
mode, thus playback start never triggers a property change for it, and
thus mpv-mpris remains in a stopped state. This temporary incorrectness
in state is automatically corrected, however, when playback is
subsequently paused.